SettingViewModelをインタフェースと実装に分ける
